Factoring chunks out for re-use between lit and react
