+++ typescript/packages/common-charm/src/index.ts
@@ -1,1 +1,11 @@
-export { runPersistent, type Charm, addCharms, removeCharm, storage, syncCharm, charms } from "./charm.js";
+export {
+  runPersistent,
+  type Charm,
+  addCharms,
+  removeCharm,
+  storage,
+  syncCharm,
+  charms,
+} from "./charm.js";
+export { syncRecipe, saveRecipe } from "./syncRecipe.js";
+export { buildRecipe, tsToExports } from "./localBuild.js";
