+++ typescript/packages/lookslike-high-level/src/data.ts
@@ -8,100 +8,24 @@ import {
   type DocImpl,
   EntityId,
   getEntityId,
-  getRecipe,
-  getRecipeParents,
-  getRecipeSrc,
   raw,
   type ReactivityLog,
-  getRecipeSpec,
-  getRecipeName,
   Action,
   addAction,
   removeAction,
 } from "@commontools/runner";
 import * as allRecipes from "./recipes/index.js";
-import { buildRecipe } from "./localBuild.js";
 import { setIframeContextHandler } from "@commontools/iframe-sandbox";
 import { addCharms } from "@commontools/charm";
-
-// Necessary, so that suggestions are indexed.
-// import "./recipes/todo-list-as-task.jsx";
-// import "./recipes/playlist.jsx";
-
 
 export const BLOBBY_SERVER_URL =
   typeof window !== "undefined"
     ? window.location.protocol + "//" + window.location.host + "/api/storage/blobby"
     : "//api/storage/blobby";
 
-const recipesKnownToStorage = new Set<string>();
-
-export async function syncRecipe(id: string) {
-  if (getRecipe(id)) {
-    if (recipesKnownToStorage.has(id)) return;
-    const src = getRecipeSrc(id);
-    const spec = getRecipeSpec(id);
-    const parents = getRecipeParents(id);
-    if (src) saveRecipe(id, src, spec, parents);
-    return;
-  }
-
-  const response = await fetch(`${BLOBBY_SERVER_URL}/spell-${id}`);
-  let src: string;
-  let spec: string;
-  let parents: string[];
-  try {
-    const resp = await response.json();
-    src = resp.src;
-    spec = resp.spec;
-    parents = resp.parents || [];
-  } catch (e) {
-    src = await response.text();
-    spec = "";
-    parents = [];
-  }
-
-  const { recipe, errors } = await buildRecipe(src);
-  if (errors) throw new Error(errors);
-
-  const recipeId = addRecipe(recipe!, src, spec, parents);
-  if (id !== recipeId) {
-    throw new Error(`Recipe ID mismatch: ${id} !== ${recipeId}`);
-  }
-  recipesKnownToStorage.add(recipeId);
-}
-
-export async function saveRecipe(
-  id: string,
-  src: string,
-  spec?: string,
-  parents?: string[],
-  spellbookTitle?: string,
-  spellbookTags?: string[],
-) {
-  // If the recipe is already known to storage, we don't need to save it again,
-  // unless the user is trying to attach a spellbook title or tags.
-  if (recipesKnownToStorage.has(id) && !spellbookTitle) return;
-  recipesKnownToStorage.add(id);
-
-  console.log("Saving recipe", id);
-  const response = await fetch(`${BLOBBY_SERVER_URL}/spell-${id}`, {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify({
-      src,
-      recipe: JSON.parse(JSON.stringify(getRecipe(id))),
-      spec,
-      parents,
-      recipeName: getRecipeName(id),
-      spellbookTitle,
-      spellbookTags,
-    }),
-  });
-  return response.ok;
-}
+// Necessary, so that suggestions are indexed.
+// import "./recipes/todo-list-as-task.jsx";
+// import "./recipes/playlist.jsx";
 
 import smolIframe from "./recipes/smolIframe.js";
 import complexIframe from "./recipes/complexIframe.js";
